Move ScoreCardBackendMock to mocks.dart
