+++ app/test/frontend/handlers_test.dart
@@ -21,6 +21,7 @@ import '../shared/handlers_test_utils.dart';
 import '../shared/utils.dart';
 
 import 'handlers_test_utils.dart';
+import 'mocks.dart';
 import 'utils.dart';
 
 void tScopedTest(String name, Future func()) {
